feat: Add function to build project
Details: Move the code from the server-py skeleton into the correct file in /src and add a function to build the project. The build is syntax checked and logged using the logger and ci_pipeline. Each build is given a unique ID using the UUID library.

Closes issue #11.
